+++ src/client/java/tsrh/xraying/client/XrayETL.java
@@ -3,12 +3,14 @@ package tsrh.xraying.client;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
+import net.minecraft.registry.Registries;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 import tsrh.xraying.client.config.Configs;
 
+import java.util.BitSet;
 import java.util.List;
 
 import static tsrh.xraying.client.Xray.mc;
@@ -18,8 +20,25 @@ public class XrayETL {
     // Load Config as status
     public static boolean isXrayActive = Configs.Generic.XRAY.getBooleanValue();
 
+    public static boolean exposedOnly =  Configs.Generic.EXPOSED_ONLY.getBooleanValue();
+
     public static int alphaWhitelist = Configs.Generic.XRAY_ALPHA.getIntegerValue();
     public static int alphaBlacklist = Configs.Generic.OTHER_ALPHA.getIntegerValue();
+    public static class AlphaBE {
+        public static int // I'll use this later
+                banner  = Configs.BlockEntities.BANNER_ALPHA.getIntegerValue(),
+                bed     = Configs.BlockEntities.BED_ALPHA.getIntegerValue(),
+                bell    = Configs.BlockEntities.BELL_ALPHA.getIntegerValue(),
+                chest   = Configs.BlockEntities.CHEST_ALPHA.getIntegerValue(),
+                conduit = Configs.BlockEntities.CONDUIT_ALPHA.getIntegerValue(),
+                decpot  = Configs.BlockEntities.DECORATE_POT_ALPHA.getIntegerValue(),
+                enctab  = Configs.BlockEntities.ENCHANTING_TABLE_ALPHA.getIntegerValue(),
+                sign    = Configs.BlockEntities.SIGN_ALPHA.getIntegerValue(),
+                hsign   = Configs.BlockEntities.HSIGN_ALPHA.getIntegerValue(),
+                skull   = Configs.BlockEntities.SKULL_ALPHA.getIntegerValue(),
+                statue  = Configs.BlockEntities.STATUE_ALPHA.getIntegerValue(),
+                shulker = Configs.BlockEntities.SHULKER_ALPHA.getIntegerValue();
+    }
 
     public static boolean fullbrightXray = Configs.Generic.FULLBRIGHT_XRAY.getBooleanValue();
     public static boolean autoFullbright = Configs.Generic.AUTO_FULLBRIGHT.getBooleanValue();
@@ -29,16 +48,22 @@ public class XrayETL {
 
     public static final List<Block> ORES = List.of(Blocks.COAL_ORE, Blocks.DEEPSLATE_COAL_ORE, Blocks.IRON_ORE, Blocks.DEEPSLATE_IRON_ORE, Blocks.GOLD_ORE, Blocks.DEEPSLATE_GOLD_ORE, Blocks.LAPIS_ORE, Blocks.DEEPSLATE_LAPIS_ORE, Blocks.REDSTONE_ORE, Blocks.DEEPSLATE_REDSTONE_ORE, Blocks.DIAMOND_ORE, Blocks.DEEPSLATE_DIAMOND_ORE, Blocks.EMERALD_ORE, Blocks.DEEPSLATE_EMERALD_ORE, Blocks.COPPER_ORE, Blocks.DEEPSLATE_COPPER_ORE, Blocks.NETHER_GOLD_ORE, Blocks.NETHER_QUARTZ_ORE, Blocks.ANCIENT_DEBRIS);
 
-    public static List<Block> WHITE_LIST = ORES;
-//    public static List<Block> BLACK_LIST = null;
+    public static final BitSet whitelistAccess = new BitSet();
+    public static void setXrayList(List<Block> blocks) {
+        whitelistAccess.clear();
+        for (Block block: blocks) {
+            if (block == null) continue;
+            whitelistAccess.set(Registries.BLOCK.getRawId(block));
+        }
+    }
 
 /**
     Block is blocked when it's not in the whitelist,
     and is not exposed
  */
     public static boolean notBlocked(Block block, BlockPos blockPos) {
-        return WHITE_LIST.contains(block)
-                && (blockPos == null || !isExposed(blockPos));
+        return whitelistAccess.get(Registries.BLOCK.getRawId(block))
+                && (!exposedOnly || blockPos == null || isExposed(blockPos));
     }
     public static boolean isBlocked(Block block, BlockPos blockPos) {return !notBlocked(block, blockPos);}
 
@@ -67,9 +92,9 @@ public class XrayETL {
     }
 
     public static int getAlpha(BlockState state, BlockPos pos) {
-        if(!isXrayActive) return -1;
+        if (!isXrayActive || state == null) return -1;
 
-        if(WHITE_LIST.contains(state.getBlock())) return alphaWhitelist;
+        if (whitelistAccess.get(Registries.BLOCK.getRawId(state.getBlock()))) return alphaWhitelist;
 
         return alphaBlacklist;
     }