+++ CHANGELOG.md
@@ -1,13 +1,24 @@
-# 1.21.11+1.0.0 (2026-05-08)
+# Changelog
+
+## Unreleased (not clear how to do)
+- Make block entity rendering translucent-able
+- GUI to select some blocks from
+- Extendable profile list
+
+## 1.21.11+1.0.0 (2026-05-08)
 - Initial release of the project.
 - Basic logics
 - Cannot be uploaded to Modrinth because of section 3.3 a of Modrinth's Content Rules.
 
-# 1.21.11+1.1.0 (2026-05-10)
+## 1.21.11+1.1.0 (2026-05-10)
 - Fixed selective block rendering logic.
 - Integration of Sodium, Lithium, Vanilla and Mod Menu.
 - Removed chest and end portal rendering mixin.
 - Removed not to be used code.
 
-# 1.21.11+1.1.1 (2026-05-10)
+## 1.21.11+1.1.1 (2026-05-10)
 - Fixed fullbright
+
+## 1.21.11+1.1.2 (2026-05-17)
+- O1 hit of searching blocks
+- Obfuscated my code. Not intended. Sorry